+++ src/main/java/io/mountblue/blogApplication/controller/PostController.java
@@ -3,42 +3,56 @@ package io.mountblue.blogApplication.controller;
 import io.mountblue.blogApplication.entity.Comment;
 import io.mountblue.blogApplication.entity.Post;
 import io.mountblue.blogApplication.entity.Tag;
+import io.mountblue.blogApplication.entity.User;
 import io.mountblue.blogApplication.repository.TagRepository;
+import io.mountblue.blogApplication.repository.UserRepository;
 import io.mountblue.blogApplication.service.PostService;
 import io.mountblue.blogApplication.service.PostTagService;
+import io.mountblue.blogApplication.service.UserService;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Sort;
 import org.springframework.format.annotation.DateTimeFormat;
+import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.*;
 
+import java.security.Principal;
 import java.time.LocalDate;
 import java.time.LocalDateTime;
 import java.time.LocalTime;
 import java.time.format.DateTimeFormatter;
 import java.util.List;
+import java.util.Objects;
 import java.util.Set;
 
 @Controller
-@RequestMapping("/posts")
+@RequestMapping("/")
 public class PostController {
 
     private final PostService postService;
     private final PostTagService postTagService;
     private final TagRepository tagRepository;
+    private final UserRepository userRepository;
 
-    public PostController(PostService postService, PostTagService postTagService, TagRepository tagRepository) {
+    public PostController(PostService postService, PostTagService postTagService, TagRepository tagRepository, UserService userService, UserRepository userRepository) {
         this.postService = postService;
         this.postTagService = postTagService;
         this.tagRepository = tagRepository;
+        this.userRepository = userRepository;
     }
 
-    @GetMapping("/create")
-    public String showCreateForm(Model model) {
+    @GetMapping("/newpost")
+    public String showCreateForm(Model model,Principal principal) {
         model.addAttribute("post", new Post());
+
+        if (principal != null) {
+            String name = principal.getName(); // usually this is the username/email
+            User user = userRepository.findByUserName(name);
+            model.addAttribute("loggedInUser", user);
+        }
         return "postForm";
     }
 
@@ -47,7 +61,7 @@ public class PostController {
                           @RequestParam("newTags") String newTags){
 
         postService.createPost(post,newTags);
-        return "redirect:/posts";
+        return "redirect:/";
     }
 
 
@@ -55,18 +69,36 @@ public class PostController {
     public String addPost(@RequestParam Long id, @ModelAttribute Post post,
                           @RequestParam("newTags") String newTags){
         postService.updatePost(id,post,newTags);
-        return "redirect:/posts/" + id;
+        return "redirect:/post" + id;
     }
 
     @GetMapping("/delete/{id}")
-    public String delPost(@PathVariable Long id){
-        postService.deletePost(id);
-        return "redirect:/posts";
+    public String delPost(@PathVariable Long id,Principal principal){
+        Post post =  postService.getPost(id);
+        if (principal != null) {
+            String name = principal.getName(); // usually this is the username/email
+            User user = userRepository.findByUserName(name);
+            if(user.getUserName().equals(post.getAuthor()) ||
+                    (user.getRole().equals("ADMIN"))
+            ){
+                postService.deletePost(id);
+            }
+        }
+        return "redirect:/";
     }
 
     @GetMapping("/edit/{id}")
-    public String showEditForm(@PathVariable Long id, Model model) {
+    public String showEditForm(@PathVariable Long id, Model model,Principal principal) {
         Post post =  postService.getPost(id);
+
+        if (principal != null) {
+            String name = principal.getName();
+            User user = userRepository.findByUserName(name);
+            model.addAttribute("loggedInUser", user);
+            if((!Objects.equals(user.getRole(), "ADMIN")) && (!Objects.equals(user.getUserName(), post.getAuthor()))){
+                return "redirect:/";
+            }
+        }
         model.addAttribute("post", post);
 
         Set<Tag> allTags = postTagService.getTagsByPostId(id);
@@ -77,61 +109,48 @@ public class PostController {
         }
         String tagString = sb.toString();
         model.addAttribute("allTags", tagString);
+
+
         return "postForm";
     }
 
-    @GetMapping("/{id}")
-    public String viewPost(@PathVariable Long id, Model model) {
+    @GetMapping("/post{id}")
+    public String viewPost(@PathVariable Long id, Model model,Principal principal) {
         Post post = postService.getPost(id);
         Set<Tag> allTags = postTagService.getTagsByPostId(id);
         model.addAttribute("post", post);
         model.addAttribute("commentObj", new Comment());
         model.addAttribute("allTags", allTags);
+
+        if (principal != null) {
+            String name = principal.getName(); // usually this is the username/email
+            User user = userRepository.findByUserName(name);
+            model.addAttribute("loggedInUser", user);
+        }
         return "postDetail";
     }
 
     @GetMapping
     public String listPosts(
-            @RequestParam(defaultValue = "1") int start,               // 1-based index
+            @RequestParam(defaultValue = "1") int start,
             @RequestParam(defaultValue = "3") int limit,
             @RequestParam(defaultValue = "createdAt") String sortField,
             @RequestParam(defaultValue = "desc") String order,
-            @RequestParam(required = false) String authorName,
+            @RequestParam(required = false, name = "authorName") List<String> authorNames,
             @RequestParam(required = false) List<Long> tagId,
             @RequestParam(required = false) @DateTimeFormat(iso = DateTimeFormat.ISO.DATE) LocalDate from,
             @RequestParam(required = false) @DateTimeFormat(iso = DateTimeFormat.ISO.DATE) LocalDate to,
+            @RequestParam(required = false) String search,
+            Principal principal,
             Model model
     ) {
-        int page = (start - 1) / limit; // convert to 0-based page index
+
+        int page = (start - 1) / limit;
         Sort.Direction direction = order.equalsIgnoreCase("asc") ? Sort.Direction.ASC : Sort.Direction.DESC;
         Pageable pageable = PageRequest.of(page, limit, Sort.by(direction, sortField));
 
-        LocalDateTime fromDateTime = (from != null) ? from.atStartOfDay() : null;
-        LocalDateTime toDateTime = (to != null) ? to.atTime(LocalTime.MAX) : null;
+        Page<Post> postPage = postService.getFilteredPosts(authorNames, tagId, from, to,search, pageable);
 
-        Page<Post> postPage;
-
-        boolean hasAuthor = authorName != null && !authorName.isEmpty();
-        boolean hasTags = tagId != null && !tagId.isEmpty();
-        boolean hasFromTo = from != null || to != null;
-
-        if (hasAuthor && hasTags && hasFromTo) {
-            postPage = postService.getPostsByAuthorTagsAndDateRange(authorName, tagId, fromDateTime, toDateTime, pageable);
-        } else if (hasAuthor && hasTags) {
-            postPage = postService.getPaginatedPostsByAuthorAndTags(authorName, tagId, pageable);
-        } else if (hasAuthor && hasFromTo) {
-            postPage = postService.getPostsByAuthorAndDateRange(authorName, fromDateTime, toDateTime, pageable);
-        } else if (hasTags && hasFromTo) {
-            postPage = postService.getPostsByTagsAndDateRange(tagId, fromDateTime, toDateTime, pageable);
-        } else if (hasAuthor) {
-            postPage = postService.getPaginatedPostsByAuthor(authorName, pageable);
-        } else if (hasTags) {
-            postPage = postService.getPaginatedPostsByTags(tagId, pageable);
-        } else if (hasFromTo) {
-            postPage = postService.getPostsByDateRange(fromDateTime, toDateTime, pageable);
-        } else {
-            postPage = postService.getPaginatedPosts(pageable);
-        }
 
         List<Tag> allTags = tagRepository.findAll();
 
@@ -149,11 +168,12 @@ public class PostController {
         model.addAttribute("totalPages", postPage.getTotalPages());
         model.addAttribute("sortField", sortField);
         model.addAttribute("order", order);
-        model.addAttribute("authorName", authorName);
+        model.addAttribute("authorNames", authorNames); // updated
         model.addAttribute("allAuthors", postService.getAllDistinctAuthors());
         model.addAttribute("allTags", allTags);
         model.addAttribute("tagIds", tagId);
         model.addAttribute("tagParamString", tagParamString.toString());
+        model.addAttribute("search", search);
 
         DateTimeFormatter formatter = DateTimeFormatter.ofPattern("yyyy-MM-dd");
         String formattedFrom = (from != null) ? from.format(formatter) : null;
@@ -161,8 +181,16 @@ public class PostController {
         model.addAttribute("from", formattedFrom);
         model.addAttribute("to", formattedTo);
 
+        if (principal != null) {
+            String name = principal.getName(); // usually this is the username/email
+            User user = userRepository.findByUserName(name);
+            model.addAttribute("loggedInUser", user);
+        }
+
         return "posts";
     }
+
+
